+++ aacgmv2/deprecated.py
@@ -1,3 +1,8 @@
+# Copyright (C) 2019 NRL 
+# Author: Angeline Burrell
+# Disclaimer: This code is under the MIT license, whose details can be found at
+# the root in the LICENSE file
+#
 # -*- coding: utf-8 -*-
 """Pythonic wrappers for AACGM-V2 C functions that were depricated in the
 change from version 2.0.0 to version 2.0.2
@@ -15,11 +20,14 @@ References
 Laundal, K. M. and A. D. Richmond (2016), Magnetic Coordinate Systems, Space
  Sci. Rev., doi:10.1007/s11214-016-0275-y.
 -------------------------------------------------------------------------------
+
 """
 
 from __future__ import division, absolute_import, unicode_literals
+import datetime as dt
 import numpy as np
-import logbook as logging
+import warnings
+import aacgmv2
 
 def convert(lat, lon, alt, date=None, a2g=False, trace=False, allowtrace=False,
             badidea=False, geocentric=False):
@@ -54,7 +62,10 @@ def convert(lat, lon, alt, date=None, a2g=False, trace=False, allowtrace=False,
     lon_out : (float)
         Output longitude in degrees E
     """
-    import aacgmv2
+
+    dstr = "Deprecated routine, will be removed in version 2.6.  Recommend "
+    dstr += "using convert_latlon or convert_latlon_arr"
+    warnings.warn(dstr, category=FutureWarning)
 
     if(np.array(alt).max() > 2000 and not trace and not allowtrace and
        not badidea):
@@ -62,7 +73,6 @@ def convert(lat, lon, alt, date=None, a2g=False, trace=False, allowtrace=False,
         estr += ' must either use field-line tracing (trace=True '
         estr += 'or allowtrace=True) or indicate you know this is a bad idea'
         estr += ' (badidea=True)'
-        logging.error(estr)
         raise ValueError(estr)
 
     # construct a code from the boolian flags
@@ -73,7 +83,7 @@ def convert(lat, lon, alt, date=None, a2g=False, trace=False, allowtrace=False,
 
     # convert location
     lat_out, lon_out, _ = aacgmv2.convert_latlon_arr(lat, lon, alt, date,
-                                                     code=bit_code)
+                                                     method_code=bit_code)
 
     return lat_out, lon_out
 
@@ -108,12 +118,19 @@ def subsol(year, doy, utime):
     algorithm).
     After Fortran code by A. D. Richmond, NCAR. Translated from IDL
     by K. Laundal.
+
     """
+
+    dstr = "Deprecated routine, may be removed in future versions"
+    warnings.warn(dstr, category=FutureWarning)
+
+    # Convert from 4 digit year to 2 digit year
     yr2 = year - 2000
 
     if year >= 2101:
-        logging.error('subsol invalid after 2100. Input year is:', year)
+        aacgmv2.logger.error('subsol invalid after 2100. Input year is:', year)
 
+    # Determine if this year is a leap year
     nleap = np.floor((year - 1601) / 4)
     nleap = nleap - 99
     if year <= 1900:
@@ -123,6 +140,8 @@ def subsol(year, doy, utime):
         ncent = 3 - ncent
         nleap = nleap + ncent
 
+    # Calculate some of the coefficients needed to deterimine the mean longitude
+    # of the sun and the mean anomaly
     l_0 = -79.549 + (-0.238699 * (yr2 - 4 * nleap) + 3.08514e-2 * nleap)
     g_0 = -2.472 + (-0.2558905 * (yr2 - 4 * nleap) - 3.79617e-2 * nleap)
 
@@ -177,6 +196,10 @@ def gc2gd_lat(gc_lat):
     gd_lat : (same as input)
         Geodetic latitude in degrees N
     """
+    dstr = "Deprecated routine, may be removed in future versions"
+    warnings.warn(dstr, category=FutureWarning)
+
+    
     wgs84_e2 = 0.006694379990141317 - 1.0
     return np.rad2deg(-np.arctan(np.tan(np.deg2rad(gc_lat)) / wgs84_e2))
 
@@ -201,8 +224,10 @@ def igrf_dipole_axis(date):
     work after IGRF updates.  The dipole coefficients are interpolated to the
     date, or extrapolated if date > latest IGRF model
     """
-    import datetime as dt
-    import aacgmv2
+
+    dstr = "Deprecated routine, may be removed in future versions"
+    warnings.warn(dstr, category=FutureWarning)
+
 
     # get time in years, as float:
     year = date.year